+++ docs/changelog.md
@@ -17,6 +17,7 @@
 - Fixed download is overflowing when `.temp` file is out of range
 - Fixed memory slowly increased during download
 - Fixed download with `raw-single` format is not working.
+- Fixed error `OSError: broken data stream when reading image file` when downloading in any `pdf` formats
 
 ### Breaking changes